+++ AdminLTE/app/Http/Livewire/ModuloComunicacion/GestionComunicacion/Index.php
@@ -5,8 +5,10 @@ namespace App\Http\Livewire\ModuloComunicacion\GestionComunicacion;
 use Livewire\Component;
 use Livewire\WithPagination;
 use Livewire\WithFileUploads;
+use DB;
 use Illuminate\Support\Facades\Storage;
 use App\Models\ModuloComunicacion\Comunicacion;
+use App\Models\ModuloComunicacion\Elemento;
 use Illuminate\Http\Request;
 
 class Index extends Component
@@ -23,7 +25,9 @@ class Index extends Component
 
     public $descripcion;
     public $name;
+    public $comunicacion_id;
     public $comunicacion;
+    public $imagen_comunicacion;
     public $imagen;
     public $status;
     public $identificador;
@@ -94,6 +98,7 @@ class Index extends Component
         $imagen = $this->imagen->store('comunicacion');
 
         Comunicacion::create([
+
             'name' => $this->name,
             'descripcion' => $this->descripcion,
             'imagen' => $imagen,
@@ -104,7 +109,7 @@ class Index extends Component
 
         $this->identificador = rand();
 
-        $this->emit('alert', '!Se Agregó una Categoria con Exito¡');
+        $this->emit('alert', '!Se agregó un elemento de categoria con exito¡');
     }
 
     public function show(Comunicacion $comunicacion)
@@ -115,12 +120,21 @@ class Index extends Component
 
     public function delete(Comunicacion $comunicacion)
     {
-        $comunicacion->delete();
+        $consulta = DB::table('elementos')->where('comunicacion_id','=', $comunicacion->id)->get();
+        $contador = count($consulta);
+
+        if($contador > 0)
+        {
+            $this->emit('error', 'Este categoria de comunicación no se puede eliminar, contiene un elemento de comunicación');
+        }else{
+            $comunicacion->delete();
+            $this->emit('alert', 'Categoria de comunicacion eliminado con exito!');
+        }
     }
 
     public function edit( Comunicacion $comunicacion)
     {
-
+        $this->comunicacion_id = $comunicacion->id;
         $this->name = $comunicacion->name;
         $this->imagen_comunicacion = $comunicacion->imagen;
         $this->descripcion = $comunicacion->descripcion;
@@ -133,17 +147,21 @@ class Index extends Component
         $this->validate([
             'name' => 'required',
             'descripcion' => 'required',
-            'status' => 'required'
+            'status' => 'required',
+            'comunicacion_id' => 'required'
         ]);
 
         if ($this->imagen) {
-            Storage::delete([$this->imagen]);
-            $this->imagen = $this->imagen->store('comunicacion');
+            Storage::delete([$this->imagen_comunicacion]);
+            $this->imagen_comunicacion = $this->imagen->store('comunicacion');
         }
 
+        $comunicacion = Comunicacion::find($this->comunicacion_id);
+
         $comunicacion->update([
+
             'name' => $this->name,
-            'imagen' => $this->imagen,
+            'imagen' => $this->imagen_comunicacion,
             'descripcion' => $this->descripcion,
             'status' => $this->status
         ]);
